+++ src/app-configurator/AppConfig.ts
@@ -4,6 +4,7 @@ import License from './License'
 export default interface AppConfig {
   [ConfigProperty.Author]: string
   [ConfigProperty.Name]: string
+  [ConfigProperty.NpmPackage]: boolean
   [ConfigProperty.License]: License | null
   [ConfigProperty.TsAdvanced]: boolean
 }