+++ src/app-configurator/ConfigProperty.ts
@@ -1,6 +1,7 @@
 enum ConfigProperty {
   Author = 'author',
   Name = 'name',
+  NpmPackage = 'isNpmPackage',
   License = 'license',
   TsAdvanced = 'tsAdvanced'
 }