+++ src/app-generator/FileCopyParams.ts
@@ -1,0 +1,5 @@
+export default interface FileCopyParams {
+  from: string
+  to?: string
+  data?: Record<string, unknown>
+}